Separate file for getInfoFromNowStatus function.
